% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Workspace.R
\docType{class}
\name{Workspace}
\alias{Workspace}
\title{Workspace}
\format{\code{\link{R6Class}} object.}
\usage{
Workspace
}
\value{
Object of \code{\link{R6Class}} with methods to easy create and update an R package.
}
\description{
Workspace
}
\section{Fields}{

\describe{
\item{\code{root}}{character. Stores absolute path of working root directory.}
}}
\section{Methods}{

\describe{
  \item{\code{new(Xroot = getwd())}}{This method is used to create object of this class with \code{Xroot} as absolute path for working root directory.}

  \item{\code{folderCreate(XfolderList)}}{This method create folders (\code{dir.create}) given a list of paths (relative with respect to \code{Xroot}) by \code{folderList} parameter.}
  \item{\code{fileCreate(XfileList)}}{This method create files (\code{file.create}) given a list of paths (relative with respect to \code{Xroot}) by \code{fileList} parameter.}
  \item{\code{fileList(Xfolder='', Xpattern = '*', Xfull = TRUE, Xrecursive = FALSE)}}{This method returns a vector of relative paths (if \code{Xfull = FALSE} return filenames). Regex filters can be applied by \code{Xpattern} argument.}
  \item{\code{fileCopy(XfileList, Xfolder)}}{This method copy a list of files into a folder. All paths relative to \code{self$root}.}
  \item{\code{gitInit(Xfolder)}}{This method initialize (init && add --all && commit -m "Init") a git repository in a given \code{Xfolder}.}
  \item{\code{gitCommit(Xfolder, Xmessage)}}{The method commit (add . && commit -m "\code{Xmessage}") the git repository in a given \code{Xfolder}.}
  \item{\code{libraryCreate(Xlibrary)}}{This method creates skeleton package with \code{create()} function from \code{roxygen2}, adds forlder to complete the package skeleton and runs \code{gitInit()}.}
  \item{\code{libraryUpdate(Xlibrary)}}{This method updates package documentation (with \code{document()} function from \code{roxygen2}), creates a pfd manual, builds and installs the package and runs \code{gitCommit()}.}
 \item{\code{readAllFiles(Xpath, Xpattern = '*.csv', Xbind = TRUE, X2env = TRUE)}}{This method import all files in a folder and returns a only \code{data.table}. If \code{Xbind = FALSE} returns a list of \code{data.table} with the same name than files and save in Global Environment. If \code{X2env = FALSE} returns a named list of \code{data.table}.}
 \item{\code{rd2pdf(Xtitle, Xoutput, Xinput)}}{This method coverts an \code{Xinput} Rd file into \code{Xoutput} pdf with \code{Xtitle} from \code{Xroot}.}   
 \item{\code{runPython(Xscript)}}{This method run python script saved in \code{Xscript} path from \code{self$root}.}
  \item{\code{runSystem(Xcommand)}}{This method run bash command typed in \code{Xcommand} from \code{self$root}.}
  }
}
\examples{
# Instance at current working directory
w = Workspace$new(getwd())
# Create a package skeleton (roxygen2::create()),
# git init, copy some files,
# set documentation (devtools::document()),
# git commit and install
name = 'newLibrary'
# Create a Workspace object at working root directory (\\code{getwd()})
w = Workspace$new()
# Create a library skeleton and git init
w$libraryCreate(name)
# Generate Rd documentation, pdf manual, build, install and git commit
w$libraryUpdate(name, 'version 0.0.0.9')

}
\author{
Joan Dídac Viana Fons \email{joanvianafons@gmail.com}
}
\keyword{Workspace,}
\keyword{git,}
\keyword{library,}
\keyword{package}
\keyword{pvR,}

